Add support for Multi Select Bottom Sheet.
