+++ src/components/ui/button.tsx
@@ -8,7 +8,7 @@ import { cx } from '../../lib/cx.tsx';
  * @see https://ui.shadcn.com/docs/components/button
  */
 const buttonVariants = cva(
-  'inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none disabled:pointer-events-none disabled:opacity-50',
+  'flex-row items-center justify-center whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none disabled:pointer-events-none disabled:opacity-50',
   {
     defaultVariants: {
       size: 'default',
